manage Cppcheck process life cycle better
This commit enhances the management of the Cppcheck process within `CppcheckExtension`. Firstly, it introduces a new class member `#cppcheckProcess` to keep track of the spawned child process for Cppcheck. This enables cancelling of previous checks before starting new ones.

In addition, the cancellation handling is improved. The code that removed the `#cancellationTokenSource` from the context subscriptions and disposed it during the reset operation has been removed.

Also, any time the Cppcheck process ends (whether due to normal completion or cancellation), it ensures that `#cppcheckProcess` is set to `undefined`, which helps prevent potential issues with stale references.

This should lead to more controlled execution of background checking processes and more predictable behavior regarding cancellation of checks.